+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Fixed
+- version detection on Linux when Unity was distributed as a TeamCity tool
 
 ## 1.0.2 - 2023-03-01